clean type index
